fix for no Go source files on macOS
Make things a bit more easier on macOS
```
~/go/src/github.com/google/cadvisor
$ go list all
build k8s.io/utils/inotify: cannot load k8s.io/utils/inotify: no Go source files
```
